✅ [test]: QueueTest구현
*Queue 프로퍼티 count 로직 수정
*test메서드 이름 수정
*Edge case test 추가
